Update ShpfyAuthenticationMgt.Codeunit.al Proposal
Addling events before getting the secret values. These events are only triggered in a sandbox or on a docker environment, not on a production environment.
This give us the possibility to run the standard version of the connector in a docker environment with own Shopify app.
This is now not possible or we must download the source code and change it every time there is a new version of it.